implement Feature #15 . change test framework from karma to jest.
