+++ client/web/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+   globals: {
+      'ts-jest': {
+         tsConfigFile: './tsconfig.spec.json',
+      },
+   },
+   setupFiles: ["jest-canvas-mock"]
+};
