+++ client/web/package.json
@@ -10,13 +10,11 @@
     "build:prod": "npm run build:ng:prod",
     "build:ng:prod": "npm run build:ng:app -- --prod --aot",
     "build:watch": "concurrently -r \"npm run build:ng:app -- --watch\"",
-    "test": "npm run test:app -- --watch false",
-    "test:app": "npm run ng -- test angboot-web",
     "lint": "tsc -p rules/tsconfig.json && npm run ng -- lint",
     "e2e": "npm run ng -- e2e --serve false",
     "generate-index-files": "node ./generate-index-files.js",
-    "jest": "jest",
-    "jest:watch": "jest --watch",
+    "test": "JEST_JUNIT_OUTPUT=./build/test-results/junit.xml npm run ng -- test angboot-web --ci --runInBand --reporters=default --reporters=jest-junit",
+    "test:watch": "jest --watch",
     "report-coverage": "codecov --token=1eb18219-afec-47c9-8f90-1210b4ec4d07"
   },
   "private": true,
@@ -43,9 +41,11 @@
     "rxjs": "^6.5.3",
     "split.js": "^1.5.11",
     "stompjs": "^2.3.3",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.9.1",
+    "jest-codemods": "^0.22.1"
   },
   "devDependencies": {
+    "@angular-builders/jest": "^8.3.2",
     "@angular-devkit/build-angular": "~0.803.20",
     "@angular/cli": "~8.3.20",
     "@angular/compiler-cli": "^8.2.14",
@@ -53,6 +53,7 @@
     "@types/geojson": "^1.0.0",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "^2.0.2",
+    "@types/jest": "^24.9.1",
     "@types/jquery": "^2.0.49",
     "@types/node": "^6.0.101",
     "clean-webpack-plugin": "1.0.1",
@@ -61,12 +62,11 @@
     "concurrently": "^3.5.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
+    "jest": "^24.9.0",
+    "jest-canvas-mock": "^2.2.0",
+    "jest-junit": "^6.3.0",
     "js-yaml": "^3.13.1",
-    "karma": "~3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
     "karma-coverage": "^1.1.2",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.2",
     "rxjs-tslint": "^0.1.7",
     "through2": "^2.0.3",
@@ -74,9 +74,5 @@
     "tslint": "~5.11.0",
     "typescript": "3.5.3",
     "vinyl-fs": "^2.4.4"
-  },
-  "jest": {
-    "preset": "jest-preset-angular",
-    "setupFilesAfterEnv": ["./setupJest.ts"]
   }
 }
