--- client/web/setupJest.ts
@@ -1,2 +1,0 @@
-import 'jest-preset-angular'; // jest 对于 angular 的预配置
-import './jestGlobalMocks'; // jest 全局的 mock
