+++ client/web/src/app/app.component.spec.ts
@@ -15,8 +15,6 @@
 import { TestBed, async } from "@angular/core/testing";
 import { AppComponent } from "./app.component";
 import { BrowserModule } from "@angular/platform-browser";
-import { Router } from "@angular/router";
-import { Subject } from "rxjs";
 import { RouterTestingModule } from "@angular/router/testing";
 
 describe("AppComponent", () => {
@@ -24,9 +22,6 @@ describe("AppComponent", () => {
     let router: any;
 
   beforeEach(async(() => {
-    router = jasmine.createSpyObj("Router", ["navigate"]);
-    router.events = new Subject<any>().asObservable();
-
     TestBed.configureTestingModule({
       imports: [
         BrowserModule,
@@ -36,10 +31,6 @@ describe("AppComponent", () => {
         AppComponent
       ],
       providers: [
-        {
-          provide: Router,
-          useValue: router
-        }
       ]
     }).compileComponents();
   }));
