+++ client/web/src/app/widget/notification/notifications.component.spec.ts
@@ -12,7 +12,6 @@
  * person.
  */
 
-import { NoopAnimationsModule } from "@angular/platform-browser/animations";
 import { TestBed, async, ComponentFixture } from "@angular/core/testing";
 import { NgbModule } from "@ng-bootstrap/ng-bootstrap";
 import { NotificationsComponent } from "./notifications.component";
