+++ client/web/src/jestGlobalMocks.ts
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2019, AngBoot Technology Corp, All Rights Reserved.
+ *
+ * The software and information contained herein are copyrighted and
+ * proprietary to AngBoot Technology Corp. This software is furnished
+ * pursuant to a written license agreement and may be used, copied,
+ * transmitted, and stored only in accordance with the terms of such
+ * license and with the inclusion of the above copyright notice. Please
+ * refer to the file "COPYRIGHT" for further copyright and licensing
+ * information. This software and information or any other copies
+ * thereof may not be provided or otherwise made available to any other
+ * person.
+ */
+
+const mock = () => {
+    let storage = {};
+    return {
+        getItem: key => key in storage ? storage[key] : null,
+        setItem: (key, value) => storage[key] = value || "",
+        removeItem: key => delete storage[key],
+        clear: () => storage = {},
+    };
+};
+
+Object.defineProperty(window, "localStorage", {value: mock()});
+Object.defineProperty(window, "sessionStorage", {value: mock()});
+Object.defineProperty(window, "getComputedStyle", {
+    value: () => ["-webkit-appearance"]
+});
