+++ client/web/tsconfig.json
@@ -14,9 +14,16 @@
     "typeRoots": [
       "node_modules/@types"
     ],
+    "types": [
+      "jest"
+    ],
     "lib": [
       "es2018",
       "dom"
     ]
+  },
+  "angularCompilerOptions": {
+    "fullTemplateTypeCheck": true,
+    "strictInjectionParameters": true
   }
 }
